v9: adjust the scheduler and the learning rate; add vector quantization; change to bfloat16

+++ multicontrast/engine/validator.py
@@ -57,7 +57,7 @@ class SupervisedValidator(BaseValidator):
         selected_contrasts = batch['selected_contrasts']
         generated_contrats = batch['generated_contrasts']
         with torch.no_grad():
-            with autocast(dtype=torch.float16):
+            with autocast(dtype=torch.bfloat16):
                 pred = self.model(x,
                                   selected_contrasts,
                                   generated_contrats,
